fix dtype

+++ delensalot/core/iterator/iteration_handler.py
@@ -65,7 +65,7 @@ class base_iterator():
                 return ret
         else:
             if self.k in ['p_p', 'p_eb', 'peb', 'p_be', 'pee']:
-                return np.array(self.sims_MAP.get_sim_pmap(self.simidx))
+                return np.array(self.sims_MAP.get_sim_pmap(self.simidx), dtype=float)
             else:
                 assert 0, 'implement if needed'
